Move toolbar to own namespace.
