+++ osu.Game/Overlays/Toolbar/ToolbarModeButton.cs
@@ -3,13 +3,9 @@
 
 using osu.Framework.Extensions;
 using osu.Game.Graphics;
-using OpenTK.Graphics;
-using osu.Framework;
-using osu.Framework.Allocation;
 using osu.Game.Modes;
-using osu.Game.Screens.Play;
 
-namespace osu.Game.Overlays
+namespace osu.Game.Overlays.Toolbar
 {
     public class ToolbarModeButton : ToolbarButton
     {
