+++ osu.Game/Overlays/Toolbar/ToolbarModeSelector.cs
@@ -3,19 +3,16 @@
 
 using System;
 using System.Linq;
+using osu.Framework.Caching;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Transformations;
+using osu.Game.Modes;
 using OpenTK;
 using OpenTK.Graphics;
-using osu.Framework;
-using osu.Framework.Caching;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Allocation;
-using osu.Game.Modes;
-using osu.Game.Screens.Play;
 
-namespace osu.Game.Overlays
+namespace osu.Game.Overlays.Toolbar
 {
     class ToolbarModeSelector : Container
     {
